+++ RedCorp_Application/src/main/java/pe/edu/upc/aaw/redcorp_application/serviceinterfaces/IUsuarioService.java
@@ -8,7 +8,7 @@ public interface IUsuarioService {
     public void insert(Usuario usuario);
     public List<Usuario> list();
 
-    public void delete(int idUsuario );
+    public void delete(Long idUsuario );
 
-    public Usuario listId(int idUsuario);
+    public Usuario listId(Long idUsuario);
 }
